Removed unwanted files delete_connector.py
