+++ stix_shifter_modules/onelogin/entry_point.py
@@ -1,9 +1,4 @@
 from stix_shifter_utils.utils.base_entry_point import BaseEntryPoint
-from stix_shifter_utils.modules.base.stix_transmission.base_sync_connector import BaseSyncConnector
-from .stix_transmission.ping_connector import PingConnector
-from .stix_transmission.delete_connector import DeleteConnector
-from .stix_transmission.results_connector import ResultsConnector
-from .stix_transmission.api_client import APIClient
 
 
 class EntryPoint(BaseEntryPoint):
@@ -13,18 +8,6 @@ class EntryPoint(BaseEntryPoint):
         self.set_async(False)
 
         if connection:
-            api_client = APIClient(connection, configuration)
-            base_sync_connector = BaseSyncConnector()
-            ping_connector = PingConnector(api_client)
-            query_connector = base_sync_connector
-            status_connector = base_sync_connector
-            results_connector = ResultsConnector(api_client)
-            delete_connector = DeleteConnector(api_client)
-
-            self.set_results_connector(results_connector)
-            self.set_status_connector(status_connector)
-            self.set_delete_connector(delete_connector)
-            self.set_query_connector(query_connector)
-            self.set_ping_connector(ping_connector)
+            self.setup_transmission_basic(connection, configuration)
 
         self.setup_translation_simple(dialect_default='default')
