+++ stix_shifter_modules/onelogin/stix_transmission/connector.py
@@ -1,22 +1,38 @@
 import json
-
-from stix_shifter_utils.modules.base.stix_transmission.base_results_connector import BaseResultsConnector
+from stix_shifter_utils.modules.base.stix_transmission.base_sync_connector import BaseSyncConnector
+from .api_client import APIClient
 from stix_shifter_utils.utils.error_response import ErrorResponder
 from stix_shifter_utils.utils import logger
 
 
-class ResultsConnector(BaseResultsConnector):
+class Connector(BaseSyncConnector):
     max_limit = 50
 
-    def __init__(self, api_client):
+    def __init__(self, connection, configuration):
         self.init_error = None
-        self.api_client = api_client
+        self.api_client = APIClient(connection, configuration)
         self.logger = logger.set_logger(__name__)
         self.token_resp = self.get_token()
         if self.token_resp["code"] == 200:
             self.access_token = self.token_resp["access_token"]
         else:
             self.init_error = self.token_resp
+
+    def ping_connection(self):
+        try:
+            response = self.api_client.generate_token()
+            response_code = response.code
+            response_dict = json.loads(response.read())
+            # Construct a response object
+            return_obj = dict()
+            if response_code == 200:
+                return_obj['success'] = True
+            else:
+                ErrorResponder.fill_error(return_obj, response_dict, ['status', 'message'])
+            return return_obj
+        except Exception as err:
+            self.logger.error('error when pinging datasource {}:'.format(err))
+            raise
 
     def create_results_connection(self, quary_expr, offset, length):
         length = int(length)
@@ -29,7 +45,7 @@ class ResultsConnector(BaseResultsConnector):
                 self.logger.error(f"Token Generation Failed: {self.init_error}")
                 return self.init_error
             # Separate out api supported url params
-            quary_expr, filter_attr = ResultsConnector.modify_query_expr(quary_expr)
+            quary_expr, filter_attr = Connector.modify_query_expr(quary_expr)
             # Grab the response, extract the response code, and convert it to readable json
             if length <= self.max_limit:
                 # $(offset) param not included as data source not support this
@@ -54,7 +70,7 @@ class ResultsConnector(BaseResultsConnector):
                             if filter_attr:
                                 filter_flag = False
                                 # filter data based on filter-attributes
-                                response_dict = ResultsConnector.filter_response(response_dict, filter_attr)
+                                response_dict = Connector.filter_response(response_dict, filter_attr)
                             return_obj['data'].extend(response_dict['data'])
                         else:
                             ErrorResponder.fill_error(return_obj, response_dict, ['status', 'message', 'description'])
@@ -62,7 +78,7 @@ class ResultsConnector(BaseResultsConnector):
                         break
                 # filter data if not filtered in above while loop
                 if filter_flag:
-                    return_obj = ResultsConnector.filter_response(return_obj, filter_attr)
+                    return_obj = Connector.filter_response(return_obj, filter_attr)
                 # slice the records as per the provided offset and length(limit)
                 return_obj['data'] = return_obj['data'][offset:total_records]
 
@@ -113,4 +129,3 @@ class ResultsConnector(BaseResultsConnector):
         except KeyError as ex:
             raise KeyError(f"Invalid parameter {ex}")
         return response_dict
-
