--- stix_shifter_modules/onelogin/stix_transmission/ping_connector.py
@@ -1,26 +1,0 @@
-from stix_shifter_utils.modules.base.stix_transmission.base_ping_connector import BasePingConnector
-from stix_shifter_utils.utils.error_response import ErrorResponder
-from stix_shifter_utils.utils import logger
-import json
-
-
-class PingConnector(BasePingConnector):
-    def __init__(self, api_client):
-        self.api_client = api_client
-        self.logger = logger.set_logger(__name__)
-
-    def ping_connection(self):
-        try:
-            response = self.api_client.generate_token()
-            response_code = response.code
-            response_dict = json.loads(response.read())
-            # Construct a response object
-            return_obj = dict()
-            if response_code == 200:
-                return_obj['success'] = True
-            else:
-                ErrorResponder.fill_error(return_obj, response_dict, ['status', 'message'])
-            return return_obj
-        except Exception as err:
-            self.logger.error('error when pinging datasource {}:'.format(err))
-            raise
